CSAFE-23695:Update object-file filter for legalholds. Objects can be filter by minimum and maximum number of legalhold they have. CSAFE-23694: Change retention field data type to int.

+++ og-object/src/main/java/com/ibm/og/object/ObjectMetadata.java
@@ -43,7 +43,7 @@ public interface ObjectMetadata extends Comparable<ObjectMetadata> {
    *  Gets the retention period
    *  @return the retention period
    */
-  long getRetention();
+  int getRetention();
 
   /**
    * Convert this instance's internal representation into bytes
